+++ consul-lambda/consul-lambda-registrator/environment.go
@@ -48,6 +48,11 @@ type Config struct {
 	// ExtensionDataPrefix is the path in Parameter Store where extension data will be written.
 	ExtensionDataPrefix string `envconfig:"CONSUL_EXTENSION_DATA_PREFIX"`
 
+	// ExtensionDataTier is the tier to use for storing data in Parameter Store.
+	// Refer to the Parameter Store documentation for applicable values.
+	// If this value is not set the default value from the AWS SDK will be used.
+	ExtensionDataTier string `envconfig:"CONSUL_EXTENSION_DATA_TIER"`
+
 	// PageSize is the maximum number of Lambda functions per page when querying the Lambda API.
 	PageSize int `envconfig:"PAGE_SIZE" default:"50"`
 }
@@ -119,7 +124,7 @@ func SetupEnvironment(ctx context.Context) (Environment, error) {
 		return env, err
 	}
 
-	env.Store = client.NewSSM(&sdkConfig)
+	env.Store = client.NewSSM(&sdkConfig, env.ExtensionDataTier)
 	env.Lambda = NewLambdaClient(&sdkConfig, env.PageSize)
 
 	err = setConsulHTTPToken(ctx, env.Store, env.ConsulHTTPTokenPath)