+++ test/acceptance/tests/basic_test.go
@@ -20,9 +20,10 @@ import (
 	"github.com/gruntwork-io/terratest/modules/terraform"
 	"github.com/hashicorp/consul/api"
 	"github.com/hashicorp/consul/sdk/testutil/retry"
+	"github.com/stretchr/testify/require"
+
 	"github.com/hashicorp/terraform-aws-consul-lambda/test/acceptance/framework/config"
 	"github.com/hashicorp/terraform-aws-consul-lambda/test/acceptance/framework/logger"
-	"github.com/stretchr/testify/require"
 )
 
 type SetupConfig struct {
@@ -30,9 +31,12 @@ type SetupConfig struct {
 }
 
 func TestBasic(t *testing.T) {
+
 	cases := map[string]struct {
-		secure     bool
-		enterprise bool
+		secure                 bool
+		enterprise             bool
+		autoPublishRegistrator bool
+		privateEcrRepoName     string
 	}{
 		"secure": {
 			secure: true,
@@ -44,6 +48,15 @@ func TestBasic(t *testing.T) {
 			secure:     true,
 			enterprise: true,
 		},
+		"secure auto publish": {
+			secure:                 true,
+			autoPublishRegistrator: true,
+		},
+		"secure auto publish with privateEcrRepoName": {
+			secure:                 true,
+			autoPublishRegistrator: true,
+			privateEcrRepoName:     fmt.Sprintf("test-ecr-repo-%s", strings.ToLower(random.UniqueId())),
+		},
 	}
 
 	for name, c := range cases {
@@ -51,11 +64,11 @@ func TestBasic(t *testing.T) {
 			config := suite.Config()
 			tfVars := config.TFVars()
 			tfVars["secure"] = c.secure
+			tfVars["arch"] = config.Arch
 			namespace := ""
 			partition := ""
 			queryString := ""
-			tfVars["consul_image"] = "public.ecr.aws/hashicorp/consul:1.15.1"
-
+			tfVars["consul_image"] = "public.ecr.aws/hashicorp/consul:1.16.1"
 			if c.enterprise {
 				tfVars["consul_license"] = os.Getenv("CONSUL_LICENSE")
 				require.NotEmpty(t, tfVars["consul_license"], "CONSUL_LICENSE environment variable is required for enterprise tests")
@@ -64,13 +77,23 @@ func TestBasic(t *testing.T) {
 				tfVars["consul_namespace"] = namespace
 				tfVars["consul_partition"] = partition
 				queryString = fmt.Sprintf("?partition=%s&ns=%s", partition, namespace)
-				tfVars["consul_image"] = "public.ecr.aws/hashicorp/consul-enterprise:1.15.1-ent"
+				tfVars["consul_image"] = "public.ecr.aws/hashicorp/consul-enterprise:1.16.1-ent"
 			}
 
 			setupSuffix := tfVars["suffix"]
 			suffix := strings.ToLower(random.UniqueId())
 			tfVars["suffix"] = suffix
 			tfVars["setup_suffix"] = setupSuffix
+
+			var setupCfg SetupConfig
+
+			if c.autoPublishRegistrator {
+				tfVars["enable_auto_publish_ecr_image"] = true
+				tfVars["consul_lambda_registrator_image"] = config.ECRImageURI
+				if c.privateEcrRepoName != "" {
+					tfVars["private_ecr_repo_name"] = c.privateEcrRepoName
+				}
+			}
 
 			setupTerraformOptions := terraform.WithDefaultRetryableErrors(t, &terraform.Options{
 				TerraformDir: "./setup",
@@ -88,7 +111,6 @@ func TestBasic(t *testing.T) {
 
 			terraform.InitAndApply(t, setupTerraformOptions)
 
-			var setupCfg SetupConfig
 			require.NoError(t, UnmarshalTF("./setup", &setupCfg))
 
 			clientServiceName := fmt.Sprintf("test_client_%s", suffix)
@@ -99,8 +121,9 @@ func TestBasic(t *testing.T) {
 			lambdaToMeshServiceName := fmt.Sprintf("lambda_to_mesh_example_%s", suffix)
 
 			var consulServerTaskARN string
+			testingT := t
 			retry.RunWith(&retry.Timer{Timeout: 3 * time.Minute, Wait: 10 * time.Second}, t, func(r *retry.R) {
-				taskListOut, err := shell.RunCommandAndGetOutputE(t, shell.Command{
+				taskListOut, err := shell.RunCommandAndGetOutputE(testingT, shell.Command{
 					Command: "aws",
 					Args: []string{
 						"ecs",
@@ -134,7 +157,7 @@ func TestBasic(t *testing.T) {
 			retry.RunWith(&retry.Timer{Timeout: 5 * time.Minute, Wait: 10 * time.Second}, t, func(r *retry.R) {
 				var services []api.CatalogService
 				err := ExecuteRemoteCommandJSON(
-					t,
+					testingT,
 					suite.Config(),
 					consulServerTaskARN,
 					"consul-server",
@@ -163,6 +186,7 @@ func TestBasic(t *testing.T) {
 					"tags":   tags,
 					"name":   meshToLambdaServiceName,
 					"region": config.Region,
+					"arch":   config.Arch,
 				},
 			})
 
@@ -218,6 +242,7 @@ func TestBasic(t *testing.T) {
 					"region": config.Region,
 					"env":    env,
 					"layers": []string{suite.Config().ExtensionARN},
+					"arch":   config.Arch,
 				},
 			})
 
@@ -255,14 +280,14 @@ func TestBasic(t *testing.T) {
 			}
 
 			for _, c := range lambdas {
-				retry.RunWith(&retry.Timer{Timeout: 60 * time.Second, Wait: 5 * time.Second}, t, func(r *retry.R) {
+				retry.RunWith(&retry.Timer{Timeout: 120 * time.Second, Wait: 5 * time.Second}, t, func(r *retry.R) {
 					var services []api.CatalogService
 					qs := queryString
 					if c.inDefaultPartition {
 						qs = ""
 					}
 					err := ExecuteRemoteCommandJSON(
-						t,
+						testingT,
 						suite.Config(),
 						consulServerTaskARN,
 						"consul-server",
@@ -278,7 +303,7 @@ func TestBasic(t *testing.T) {
 				// Create an intention to allow the Lambda function to call the test_client service
 				retry.RunWith(&retry.Timer{Timeout: 60 * time.Second, Wait: 5 * time.Second}, t, func(r *retry.R) {
 					result, err := ExecuteRemoteCommand(
-						t,
+						testingT,
 						suite.Config(),
 						consulServerTaskARN,
 						"consul-server",
@@ -295,7 +320,7 @@ func TestBasic(t *testing.T) {
 				// Export the test_client service so it can be called by the Lambda function.
 				retry.RunWith(&retry.Timer{Timeout: 60 * time.Second, Wait: 5 * time.Second}, t, func(r *retry.R) {
 					result, err := ExecuteRemoteCommand(
-						t,
+						testingT,
 						suite.Config(),
 						consulServerTaskARN,
 						"consul-server",
@@ -317,7 +342,7 @@ func TestBasic(t *testing.T) {
 			defer os.Remove(outFile.Name())
 
 			retry.RunWith(&retry.Timer{Timeout: 120 * time.Second, Wait: 5 * time.Second}, t, func(r *retry.R) {
-				_, err := shell.RunCommandAndGetOutputE(t, shell.Command{
+				_, err := shell.RunCommandAndGetOutputE(testingT, shell.Command{
 					Command: "aws",
 					Args: []string{
 						"lambda",
@@ -364,7 +389,7 @@ func TestBasic(t *testing.T) {
 			retry.RunWith(&retry.Timer{Timeout: 60 * time.Second, Wait: 5 * time.Second}, t, func(r *retry.R) {
 				var services []api.CatalogService
 				err := ExecuteRemoteCommandJSON(
-					t,
+					testingT,
 					suite.Config(),
 					consulServerTaskARN,
 					"consul-server",