+++ test/acceptance/tests/mesh-to-lambda/main.tf
@@ -5,9 +5,13 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "4.14.0"
+      version = "4.67.0"
     }
   }
+}
+
+provider "aws" {
+  region = var.region
 }
 
 provider "aws" {
@@ -23,4 +27,5 @@ module "mesh_to_lambda" {
   name   = var.name
   tags   = var.tags
   region = var.region
+  arch   = var.arch
 }