+++ test/acceptance/tests/mesh-to-lambda/variables.tf
@@ -12,3 +12,9 @@ variable "name" {
 variable "region" {
   type = string
 }
+
+variable "arch" {
+  type        = string
+  default     = "x86_64"
+  description = "Lambda Architecture"
+}