+++ CHANGELOG.md
@@ -1,10 +1,22 @@
+## 0.1.0-beta5 (Mar 04, 2024)
+
+FEATURES
+* Add support for storing parameter values greater than 4 KB. The `lambda-registrator` module and source code have been updated to accept a configurable value for the [SSM parameter tier](https://docs.aws.amazon.com/systems-manager/latest/userguide/parameter-store-advanced-parameters.html). This allows users to choose if they want to use the `Advanced` tier feature. Charges apply for the `Advanved` tier so if the tier is not expressly set to `Advanced`, then the `Standard` tier will be used. Using the `Advanced` tier allows for parameter values up to 8 KB. The Lambda-registrator Terraform module can be configured using the new `consul_extension_data_tier` variable.
+  [[GH-78]](https://github.com/hashicorp/terraform-aws-consul-lambda/pull/78)
+
+* Add support for pushing `consul-lambda-registrator` public image to private ecr repo through terraform.
+  [[GH-82]](https://github.com/hashicorp/terraform-aws-consul-lambda/pull/82)
+
+* Add  arm64 support to `consul-lambda-registrator` and `consul-lambda-extension`.
+  [[GH-90]](https://github.com/hashicorp/terraform-aws-consul-lambda/pull/90)
+
 ## 0.1.0-beta4 (Apr 28, 2023)
 
-IMPROVEMENTS:
+IMPROVEMENTS
 * Pin the version of the `terraform-aws-modules/eventbridge/aws` module to v1.17.3. This ensures the selection of the eventbridge module is deterministic when using the `lambda-registrator` Terraform module.
   [[GH-70]](https://github.com/hashicorp/terraform-aws-consul-lambda/pull/70)
 
-BUG FIXES:
+BUG FIXES
 * Disable Cgo compilation for Lambda registrator and extension. Compiling without `CGO_ENABLED=0` on Go 1.20 [causes an issue](https://github.com/hashicorp/terraform-aws-consul-lambda/issues/57) that does not allow Lambda registrator or the Lambda extension to execute within the AWS Lambda runtime.
   [[GH-68]](https://github.com/hashicorp/terraform-aws-consul-lambda/pull/68)
 
@@ -19,7 +31,7 @@ BREAKING CHANGES
 FEATURES
 * Update minimum go version for project to 1.20 [[GH-1908]](https://github.com/hashicorp/terraform-aws-consul-lambda/pull/54)
 
-BUG FIXES:
+BUG FIXES
 * Security:
     * Upgrade to use Go 1.20.1 This resolves vulnerabilities [CVE-2022-41724](https://go.dev/issue/58001) in `crypto/tls` and [CVE-2022-41723](https://go.dev/issue/57855) in `net/http`. [[GH-1908]](https://github.com/hashicorp/terraform-aws-consul-lambda/pull/54)